Extract PublicationFilter component
